MAT-15 - make retries for non-realtime data non-transactional
It appears that the EM closing is specific to transactions (https://www.kerstner.at/2014/09/doctrine-2-exception-entitymanager-closed/)
And the number of errors seemed to increase when making this transactional (though not inconceivable it was chance).
Since Doctrine data is no longer realtime retrying on a best effort basis should be OK, so let's try this without a transactional
wrapper.