Make unit test of status check action always fail to connect to Redis for now
Using the real hostname means behaviour will differ depeneding on the `docker-compose` dependency being
available or not. For now we want unit tests external-dependency-free and there's limited value in replacing
the real Redis instance with an interface just for this test. So the simplest change to make the test
non-trivial and have it pass on all envs is to set an always-failing fake Redis hostname and change the
expected output to a 500 error with the Redis failure message.