MAT-296: Create command to push info about which accounts have passwords from identity database into Stripe
Merging this won't do anything by itself, we still need to adjust the
infrastructure to supply the environment variables required, and then
run the command some number of times, either manually or as a cron job.

Once its been run enough to get through the old data we can delete it.

foo