DON-893: Wip allow confirming donation payments from server
Currently the frontend tells stripe to take the payment and stripe tells
matchbot when it has it. This change allows us to insert matchbot as an
extra step, so the frontend tells stripe when the donor is ready to pay,
matchbot tells stripe, then stripe tells matchbot when the payment is
recieved.

This should give matchbot a bit more control