MAT-321: Prevent persisting any Campaign without a Charity ID
Having a campaign without a charity might be the cause of a crash on
attempt to persist a donation. This may bring any crash earlier and make
it easier to understand deal with.

I have checked the production database, there are currently no campaigns
without charities, i.e. the following returns the empty relation:

```
   SELECT * from Campaign where Campaign.charity_id IS NULL;
```