MAT-321: Delete migration that throws error when run on staging & regression, do check in ORM callback instead
In staging and regression we get an error running this:
	 SQLSTATE[HY000]: General error: 1832 Cannot change column 'charity_id': used in a foreign key constraint 'FK_E663708BF5C97E37'

Not sure why we don't get that error in prod. Might be able to fix by
dropping foreign key refs, but doing that in prod worries me a bit so
just deleting for now.

See e.g. https://stackoverflow.com/a/48140415/2526181