MAT-322: Partial spike moving away from using Doctrine associations
We've been seeing errors with saving associated objects in Doctrine,
related to replacing the EntityManager in the middle of a request, after
some things were flushed to the first object. When it's replaced it
the new one wants to save objects found through associations, but if
we tell it to try to save them then the actual database generates a
duplicate item error.

One alternative is to have ID numbers on entities instead of references
to other entities - not allowing the ORM to do either eager or lazy
loading of related entities for us writing our own code to fetch them
when required.

Started here by replacing Donation#campaign: Campaign with Donation#campaignId: int
and fixing Psalm errors. There are a load of Psalm errors showing up
still but I think I've fixed enough to illustrate the process