MAT-306: Merge stripe cash balance update webhoook controller with general webook controller
Setting up a different webhook URI seems to requiring configuring a new
signing key. Easier I think to keep the different webhooks handled in
the same controller and just switch on the event type.

We can always delegate to a another class if we want to stop this
controller getting too complicated.

Activiting this will mean changing the config of the existing stripe
webhook instead of adding a new one.