MAT-331: Increase scope for patching incorrect donation data
The bugfix for new donations is not yet deployed, so we'll need the
patch script to include donations created today and up until the point
we fully deploy the bugfix. Hoping that will be before Thursday.

Having the scope extend a little beyond the time the fix is deployed is
not a problem.