Refactor: Rename SalesforceReadProxyRepository::pull to updateFromSf and make it return void
I think this makes the function quite a bit simpler to understand. For
the one and only case where a return value was necassary, I have created a new
function CampaignRepository::pullNewFromSf. In all other cases were are
updating an existing object that the calling function is already holding
a reference to.