MAT-330: Hard code fee calculator settings
These settings don't change frequently, and the code will become simpler
if we can hard code them instead of passing them in to the calculator.

If we are passing them in then we really should be running unit tests
with other possible settings, but that doesn't seem like a good use of
time.