MAT-358: Fix Deploying change to Doctrine Entity field name generates errors
We now store the commit ID into a PHP file, and use that as part of the
cache key for the Doctrine metadata in Redis. Therefore after any deploy
of a new version of Matchbot the metadata cache will be effectively
empty.

Writing the commit ID into a PHP file rather than a plain text file so
that its content can be cached by opcache