BG2-2632: Upgrade stripe API version to 2023-08-16
See https://docs.stripe.com/upgrades#2023-08-16

The new requirement about return_url doesn ot apply to us because in the
case of a Card based donation we set `'allow_redirects
' => 'never',` when creating the intent, and in the case of a customer
balance based donation we set the `payment_method_types`.

As described at https://docs.stripe.com/upgrades/manage-payment-methods
this means we are not required to pass a `return_url` param.