BG2-2620: Replace use of Query Builder with DQL in CampaignRepository::findRecentLiveAndPendingGiftAidApproval
DQL is easier to read and work with, and I'm planing on editing this
next. DQL string was obtained by calling \Doctrine\ORM\Query::getDQL
while running integration tests, and only whitespace parts edited. There
was no conditional logic in the query generation so the DQL will be the
same every time this runs.