Restore DB columns that have been dropped from staging but not yet dropped from prod
Starting to regret dropping these in advance of deploying the code that
doens't need them to prod. Wanting to follow the more conservative path
of first deploying the code that doesn't require these columns and only
after that dropping the columns.

When we're ready to drop the columns we can do another migration a
try-catch to drop them just in environments where they're not dropped
yet.