MAT-384: Don't recalculate donation fees after payment
Need consider both MAT-384 and the old ticket MAT-330

We want to make sure we keep the donation fees as charged in the
database, not anything recalculated after the payment.

In theory both these should be the same but in case of any errors
or future changes to charging structure priority should be given to
recording the fee we calculated to pass to stripe to control what we
actually chaged, not a later calculation of what we should have charged.

In case there's a DB error and we can't save the donation fee update we
will let the process crash without collecting and it will be up to the
donor to try again hopefully that will be very rare.

There's a new call to flush changes to DB in DonationService but this
shouldn't increase load much since it means the later flush in the
Confirm action will have little or no work to do as the Doctrine UoW
will see that there are no new changes.