DON-1082: Start auto-running donation funds expired tip cancellation code again
I think we should have done this about 20 days ago but forgot. Looks
like the reason we had to comment out this line previously was fixed in
https://github.com/thebiggive/matchbot/pull/1061 - and that already
reverted the hotfix commenting the lines out but maybe something went
wrong with the merge of that into main so these lines stayed commented
out.