Release match funds when auto-cancelling donation funds donation due to failed confirmation
It's only very rarely that we should need to cancel a donation funds
donation like this, as they will normally be auto confirmed without a
problem. It can happen if the donor tries makes multiple pending
donations side by side and uses up their match funds before confirming
them.

Wasn't a big issue as when we didn't release the match funds here they
would have been found and auto released half an hour later by
DonationRepository::findWithExpiredMatching, which is able to find
cancelled donations that still have fundingWithdrawls attached.

There is a similar issue with other places we cancel auto cancel donations,
which are in confirmPreAuthorized in case a regular giving campaign is
ended, and in StripePaymentsUpdate::handlePaymentIntentCancelled which
is generally for if Stripe cancells a donation for us because e.g. they
think the payment card may be being misused. Not a big issue in either
case for now if the match funds stay tied up for half an hour.

Also added comments