DON-1066: Refactor, start moving methods for getting donation API models into donation service
This will allow us to enrich the data with more than what is available
on the actual donation record. Specifically we need the mandate UUID to
be able to generate a link to the mandate, but we may want other data
about the mandate too.

The donation service will be able to fetch that from the mandate
repository.