DON-1066: Refactor, use InMemoryDonationRepository instead of Prophecy Mocks in UpdateTest
The amount of repetitive code that I edited in Donations/UpdateTest a
couple of commits ago started to annoy me, so I have cut a lot of it out
here but usin the InMemory repository.

It's possible that some of the expectations on the prophecies that I've
deleted had value, but IMHO not likely enough to justify the reptitive
code that was making it hard to make changes.