DON-1009: Add remaining values to regular giving confirmation email
Maybe the code for multiplying by 25% and summing values is overengineerd.
We do want to be confident that we know exactly what it does though. To some extent this
is a result of having a mix numeric-string and Money values in the
existing code, and also using a purely in-house Money VO rather than off
the shelf one e.g. from Brick

Created the sum function partly because the bcadd function that we use
elsewhere doesn't have an option to add more than two values.

Wondering if we should store total of withdrawals in the DB instead of
always summing at the time. Particularly when we're looking at an old
donation it feels like it would be more natural to be looking at a
record of something calculated on the donation day rather than doing a
sum at that point.