DON-1119: WIP allow donor to complete 3D Secure authentication for first payment in regular giving setup
This is going to need some more work to fairly substantially change
exactly when we activate the regular giving mandate. Since the user's
final click might be on the 3D secure auth button, not our own site we
need to start if based on recieving an event from Stripe, not directly
based on a POST from our own site.

I think the simplest thing is to make this change apply to all regular
giving mandates, not just the ones that need the 3DS authentication.

This might mean we need the FE to poll for mandate status to know when
to show an activated mandate to the user. Stripe is not going to block
waiting for the mandate to be setup in matchbot.