DON-995: Keep info up to date from SF for all regular giving campaigns
This should mean that any changes to regular giving campaigns (in
particular the collection end date) made in SF get pulled into matchbot
within 1/2 an hour, even if the campaign is long closed.

I'd be happier if we could make it a delay of a few seconds instead of
minutes like we now do for campaigns but I think this is good enough for
now.