MAT-400: Fix ORM crash on null embeddable property
I think I misinterpreted this in the doctrine docs:

>  In case all fields in the embeddable are nullable, you might want to initialize the embeddable,
>  to avoid getting a null value instead of the embedded object.

I thought they meant if not all fields were nullable it wouldn't be
necassary to initialize the embeddable because the ORM would recognise
when a non-nullable field was null in the DB and make the entire
embeddable null.

But I don't think it is doing anything that clever, looks like it always
constructs an instance of the embeddable when hydrating an object, so it
can never be null on an entity that's been loaded from the DB. If
there's a null in the DB for a non-nullable property then that property
doesn't get initialized and the app crashes when it tries to read it.

Don't know if would be better to not use embeddables for something like
this that needs to be semantically if not technically nullable.