DON-1003: Don't attempt to send donations to Salesforce until after the related mandate has a salesforce ID
Ideally I'd like to tell Salesforce to use the UUIds that matchbot
generates to link donations to mandates, but I don't think that's
possible. So we have to delay pushing the donations until after the
mandate has an sfid.

This does also push all donations unecassarily on any later updates to a
mandate, but I think that's not a big performance issue, and we can
optimise it if necassary.