MAT-400: Switch on sending donation thanks notifications directly not via salesforce
This is set to enable for any donations confirmed after 2.30pm today. In
principle should deploy before that, although I think it may actually be
fine to deploy at any point after that as well to turn on immediatly - I
think I was wrong to think we need to use the collectedAt date and it's
actually fine to just replace `$donation->getCollectedAt() > new
\DateTimeImmutable(Donation::MAT_400_ENABLE_TIMESTAMP` with a hard-coded
true.