Add MYSQL schema to files with commands to update files from DB and check they match
Intended to make it easier for us to see what our DB schema is by having
it as files in the repo, and to see changes as diffs.

To update files after changing DB schema content:
```shell
	./matchbot matchbot:write-schema-files
```

To check if files match:
```shell
./matchbot matchbot:write-schema-files --check
```

The latter has been added to circleCI config.

This could be used in in almost any app with a DB - similar to how Ruby
On Rails generates a schema.rb file. I'm not aware of any existing tool
to do this in PHP, but there may well be one.