MAT-411: Check campaign rendered from matchbot against what SF sent us.
In future we will probably want to generate an alarm and/or throw if
this doesn't match. Right now we know it doesn't match so we just
include the list of errors in the response, and FE can display it on the
donation page (at least outside prod) until we've fixed them all.