MAT-423: WIP implement matchFundsRemaining for campaign rendering (and campaign summary in future)
I've started by just pasting in the code from
Current_Match_Funds_Available__c and adapting the syntax from SF formula
to PHP - that's easy but it depends on quite a lot of things that we
don't have ready to call on in PHP yet. I can continue this way to get
something that should be more or less gurnateed to match what SF returns
now but maybe that's no more important than getting something that
matches the actual behaviour of matchbot.

So instead of continuing to port the code as-is from SF it might be
better to rewrite from scratch based on the existing
\MatchBot\Domain\CampaignFundingRepository::getAvailableFundings(Campaign)
function, so the amount we show as available could be guranteed to match
what we are prepared to match, as long as no-one else attempts to donate
in the meantime.

I'm leaning towards the not porting. Would be good to get more opinions
though.