MAT-405: Calculate metacampaign amount raised without using donation records from SF
We do however need to pull in any adjustment amount from SF on the
metacampaign object.

Also need to check if it's correct that we're including gift aid here
and not including gift aid on the amount raised for an individual
charity campaign.

Need to add tests before releasing this