MAT-405 – fix Metacampaign summary length
We know 255 characters was insufficient. I suspect 500 would be OK for meta-campaigns but going with 1,000 for
now to be reasonably safe.

Technically the Salesforce field supports 130,000 but it's also used by charity campaigns including imported ones,
so I don't know if we can very quickly decide a sensible lower limit or enforce one at source for meta-campaigns.
As far as I can see we don't have a dedicated equivalent to this field for charity campaigns, so can probably rely on
JSON typed fields being much longer and assume we won't have an equivalent bug for those.