DON-1136: Make sure donations can't be confirmed with less match funds than expected
When we initially allocate match funds to a donation we make a record
that the donor will now be expecting those match funds to be used. We
don't allow the donation to be confirmed with less matching than that,
unless the donor tells us that they're OK with making an unmatched
donation.