MAT-452: Adjust future regular donations for one mandate to be at expected dates instead of one month later
This is a mandate that was made after 11pm, when the dates where
different in London time and UTC, and a bug meant that we collected the
second donation very shortly after the first intead of waiting a month
as we should have. That donation was refunded but now we have to fix
the problem that all future donations would be one month late as the
refunded donation is included in the count of how many months later each
donation should be taken.