INFRA-158: Skip updating any campaigns for charities that have bad regulator numbers
Slightly relying on chance here although it probably doesn't matter: The
code was written to update the charity in memory before the campaign,
so throwing on the attempt to update the charity will mean the update to
the campaign also gets skipped. If it was the other way around we would
update the campaign but leave the charity outdated which probably
wouldn't be terrible either but doesn't match what the message says.